requirements.txt
flask==3.0.3
flask-cors==4.0.0
torch>=2.1.0
numpy>=1.26.0
pillow>=10.0.0
faiss-cpu>=1.7.4
git+https://github.com/openai/CLIP.git
ftfy>=6.1.1
regex>=2024.4.24
tqdm>=4.66.0
